# cmd/swagger/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Greet service API specification.",
        "title": "API Specification",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "BillyKore",
            "url": "https://www.swagger.io/support",
            "email": "billyimmcul2010@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api.bankyaya.co.id",
    "basePath": "/api/v1",
    "paths": {
        "/scheduler": {
            "post": {
                "description": "Create a new transaction schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create a new schedule",
                "parameters": [
                    {
                        "description": "Create a new schedule request",
                        "name": "CreateScheduleRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scheduler.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/scheduler/{id}": {
            "get": {
                "description": "Retrieve a transaction schedule by the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get a schedule by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "GetScheduleRequest",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a transaction schedule by the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Delete a schedule by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "DeleteScheduleRequest",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/transfer/inquiry": {
            "post": {
                "description": "Create new inquiry transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Transfer inquiry",
                "parameters": [
                    {
                        "description": "Inquiry request",
                        "name": "InquiryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.InquiryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/transfer/payment": {
            "post": {
                "description": "Performs transfer payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Transfer payment",
                "parameters": [
                    {
                        "description": "Payment request",
                        "name": "PaymentRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "User login to get access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Inquiry request",
                        "name": "InquiryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.InquiryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "serverTime": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "scheduler.CreateScheduleRequest": {
            "type": "object",
            "required": [
                "accountType",
                "amount",
                "bankCode",
                "destination",
                "destinationName",
                "frequency",
                "sakuId",
                "startDate",
                "transactionMethod"
            ],
            "properties": {
                "accountType": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "bankCode": {
                    "type": "string"
                },
                "biFastCode": {
                    "type": "string"
                },
                "date": {
                    "type": "integer"
                },
                "day": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "destinationName": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "purposeType": {
                    "type": "string"
                },
                "sakuId": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "transactionMethod": {
                    "type": "string"
                }
            }
        },
        "transfer.InquiryRequest": {
            "type": "object",
            "required": [
                "amount",
                "destinationAccount",
                "sourceAccount"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "destinationAccount": {
                    "type": "string"
                },
                "sourceAccount": {
                    "type": "string"
                }
            }
        },
        "transfer.PaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "destinationAccount",
                "sequence",
                "sourceAccount"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "destinationAccount": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "sequence": {
                    "type": "string"
                },
                "sourceAccount": {
                    "type": "string"
                }
            }
        }
    }
}